🔨 refactor: 컴포넌트, 페이지 레이아웃 수정
