+++ src/components/Account/Description/index.tsx
@@ -2,9 +2,9 @@ import { Wrapper } from './index.style';
 
 interface DescriptionProps {
   text: string;
-  color?: string;
+  color?: 'main' | 'error' | 'green';
 }
-function Description({ text, color }: DescriptionProps) {
+function Description({ text, color = 'main' }: DescriptionProps) {
   return <Wrapper color={color}>{text}</Wrapper>;
 }
