+++ src/components/Account/SignInBtn/index.tsx
@@ -1,10 +1,17 @@
+import { useNavigate } from 'react-router-dom';
 import Button from '@/components/Common/Button';
+import { PATH } from '@/constants/path';
 
-interface SignInBtnProps {
-  startAnonymousExperience: () => void;
+interface ButtonType {
+  onClick?: () => void;
 }
+function SignInBtn({ onClick }: ButtonType) {
+  const navigate = useNavigate();
 
-function SignInBtn({ startAnonymousExperience }: SignInBtnProps) {
+  const startAnonymousExperience = () => {
+    navigate(PATH.ROOT);
+  };
+
   return (
     <>
       <Button
@@ -14,6 +21,7 @@ function SignInBtn({ startAnonymousExperience }: SignInBtnProps) {
           height: '2.47rem',
           width: '17.5rem'
         }}
+        onClick={onClick}
       />
       <Button
         content="익명으로 체험하기"
