+++ src/pages/SignInPage/index.tsx
@@ -1,10 +1,51 @@
-import SignInForm from '@/components/Account/SignInForm';
+import { SubmitHandler, useForm } from 'react-hook-form';
+import { Link } from 'react-router-dom';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { z } from 'zod';
+import Account from '@/components/Account';
+import Description from '@/components/Account/Description';
+import FormInput from '@/components/Account/FormInput';
+import SignInBtn from '@/components/Account/SignInBtn';
+import { useSignInMutation } from '@/hooks/api/useSignInMutation';
+import { PATH } from '@/constants/path';
+import * as Style from '@/pages/SignUpPage/index.style';
+import { SignInSchema } from '@/utils/validation';
+
+export type SignInSchemaType = z.infer<typeof SignInSchema>;
 
 function SignInPage() {
+  const {
+    register,
+    handleSubmit,
+    formState: { errors }
+  } = useForm<SignInSchemaType>({ resolver: zodResolver(SignInSchema) });
+
+  const { mutateSignIn } = useSignInMutation();
+
+  const handleSignInSubmit: SubmitHandler<SignInSchemaType> = (signInData) => {
+    mutateSignIn(signInData);
+  };
+
   return (
-    <>
-      <SignInForm />
-    </>
+    <Account>
+      <Style.Form onSubmit={handleSubmit(handleSignInSubmit)}>
+        <Style.FormTitle>로그인</Style.FormTitle>
+        <FormInput
+          errors={errors}
+          register={register}
+        />
+        <Style.SignUpLink>
+          <Description text="대박사건이 처음이신가요?" />
+          <Link to={PATH.SIGNUP}>
+            <Description
+              text="회원가입"
+              color="green"
+            />
+          </Link>
+        </Style.SignUpLink>
+        <SignInBtn />
+      </Style.Form>
+    </Account>
   );
 }
